+++ src/main/java/com/thebrokenrail/modupdater/api/UpdateStrategy.java
@@ -7,4 +7,12 @@ import javax.annotation.Nullable;
 public interface UpdateStrategy {
     @Nullable
     ModUpdate run(ConfigObject obj, String oldVersion, String name);
+
+    default boolean isStrict(ConfigObject obj) {
+        try {
+            return obj.getBoolean("strict");
+        } catch (ConfigObject.MissingValueException e) {
+            return true;
+        }
+    }
 }